refactor(modules): extract tracking bounded context
Moves the MR/PR lifecycle tracking domain into src/modules/tracking/.

Entity: tracking (TrackedMr, ReviewEvent, ReviewRequestState)
Use cases: trackAssignment, recordPush, recordReviewCompletion,
           checkFollowupNeeded, transitionState, syncThreads
Routes: mrTracking.routes.ts, mrTrackingAdvanced.routes.ts
Gateways: reviewRequestTracking.gateway.ts (contract — should ideally
          live in entities/; left here for Phase 2 cleanup),
          fileSystem/reviewRequestTracking.fileSystem.ts

Normalized ~25 broken relative imports in mrTrackingAdvanced.routes.ts
(the fat controller hot spot from the DDD doc) and in tests.
