--- src/interface-adapters/gateways/reviewRequestTracking.gateway.ts
@@ -1,1 +1,0 @@
-export type { ReviewRequestTrackingGateway, Platform } from '@/entities/tracking/reviewRequestTracking.gateway.js';
