+++ src/modules/tracking/entities/tracking/assignmentInfo.ts
