+++ src/modules/tracking/entities/tracking/reviewRequestTracking.gateway.ts
@@ -1,6 +1,6 @@
-import type { MrTrackingData } from '@/entities/tracking/mrTrackingData.js';
-import type { TrackedMr } from '@/entities/tracking/trackedMr.js';
-import type { ReviewEvent } from '@/entities/tracking/reviewEvent.js';
+import type { MrTrackingData } from '@/modules/tracking/entities/tracking/mrTrackingData.js';
+import type { TrackedMr } from '@/modules/tracking/entities/tracking/trackedMr.js';
+import type { ReviewEvent } from '@/modules/tracking/entities/tracking/reviewEvent.js';
 
 export type Platform = 'gitlab' | 'github';
