+++ src/modules/tracking/interface-adapters/gateways/reviewRequestTracking.gateway.ts
@@ -1,0 +1,1 @@
+export type { ReviewRequestTrackingGateway, Platform } from '@/modules/tracking/entities/tracking/reviewRequestTracking.gateway.js';
