+++ src/modules/tracking/usecases/tracking/syncThreads.usecase.ts
@@ -1,7 +1,7 @@
-import type { UseCase } from '../../shared/foundation/usecase.base.js';
-import type { ReviewRequestTrackingGateway } from '../../interface-adapters/gateways/reviewRequestTracking.gateway.js';
+import type { UseCase } from '@/shared/foundation/usecase.base.js';
+import type { ReviewRequestTrackingGateway } from '@/modules/tracking/interface-adapters/gateways/reviewRequestTracking.gateway.js';
 import type { ThreadFetchGateway } from '@/modules/platform-integration/entities/threadFetch/threadFetch.gateway.js';
-import type { TrackedMr } from '../../entities/tracking/trackedMr.js';
+import type { TrackedMr } from '@/modules/tracking/entities/tracking/trackedMr.js';
 
 interface SyncThreadsInput {
   projectPath: string;
