+++ src/modules/tracking/usecases/tracking/trackAssignment.usecase.ts
@@ -1,7 +1,7 @@
-import type { UseCase } from '../../shared/foundation/usecase.base.js';
-import type { ReviewRequestTrackingGateway } from '../../interface-adapters/gateways/reviewRequestTracking.gateway.js';
-import type { TrackedMr } from '../../entities/tracking/trackedMr.js';
-import { createTrackedMrId } from '../../entities/tracking/trackedMr.js';
+import type { UseCase } from '@/shared/foundation/usecase.base.js';
+import type { ReviewRequestTrackingGateway } from '@/modules/tracking/interface-adapters/gateways/reviewRequestTracking.gateway.js';
+import type { TrackedMr } from '@/modules/tracking/entities/tracking/trackedMr.js';
+import { createTrackedMrId } from '@/modules/tracking/entities/tracking/trackedMr.js';
 
 interface TrackAssignmentInput {
   projectPath: string;
