+++ src/modules/tracking/usecases/tracking/transitionState.usecase.ts
@@ -1,6 +1,6 @@
-import type { UseCase } from '../../shared/foundation/usecase.base.js';
-import type { ReviewRequestTrackingGateway } from '../../interface-adapters/gateways/reviewRequestTracking.gateway.js';
-import type { TrackedMr } from '../../entities/tracking/trackedMr.js';
+import type { UseCase } from '@/shared/foundation/usecase.base.js';
+import type { ReviewRequestTrackingGateway } from '@/modules/tracking/interface-adapters/gateways/reviewRequestTracking.gateway.js';
+import type { TrackedMr } from '@/modules/tracking/entities/tracking/trackedMr.js';
 
 interface TransitionStateInput {
   projectPath: string;
