+++ src/tests/units/interface-adapters/controllers/webhook/gitlab.controller.test.ts
@@ -84,13 +84,13 @@ import { invokeClaudeReview } from '@/claude/invoker.js';
 import { GitLabEventFactory } from '@/tests/factories/gitLabEvent.factory.js';
 import { createStubLogger } from '@/tests/stubs/logger.stub.js';
 import { TrackedMrFactory } from '@/tests/factories/trackedMr.factory.js';
-import type { TrackedMr } from '@/entities/tracking/trackedMr.js';
-import { TrackAssignmentUseCase } from '@/usecases/tracking/trackAssignment.usecase.js';
-import { RecordReviewCompletionUseCase } from '@/usecases/tracking/recordReviewCompletion.usecase.js';
-import { RecordPushUseCase } from '@/usecases/tracking/recordPush.usecase.js';
-import { TransitionStateUseCase } from '@/usecases/tracking/transitionState.usecase.js';
-import { CheckFollowupNeededUseCase } from '@/usecases/tracking/checkFollowupNeeded.usecase.js';
-import { SyncThreadsUseCase } from '@/usecases/tracking/syncThreads.usecase.js';
+import type { TrackedMr } from '@/modules/tracking/entities/tracking/trackedMr.js';
+import { TrackAssignmentUseCase } from '@/modules/tracking/usecases/tracking/trackAssignment.usecase.js';
+import { RecordReviewCompletionUseCase } from '@/modules/tracking/usecases/tracking/recordReviewCompletion.usecase.js';
+import { RecordPushUseCase } from '@/modules/tracking/usecases/tracking/recordPush.usecase.js';
+import { TransitionStateUseCase } from '@/modules/tracking/usecases/tracking/transitionState.usecase.js';
+import { CheckFollowupNeededUseCase } from '@/modules/tracking/usecases/tracking/checkFollowupNeeded.usecase.js';
+import { SyncThreadsUseCase } from '@/modules/tracking/usecases/tracking/syncThreads.usecase.js';
 
 function createMockTrackingGateway() {
   const basicMr = TrackedMrFactory.create({
