+++ src/usecases/triggerReview.usecase.ts
@@ -1,6 +1,6 @@
 import type { Logger } from 'pino';
 import type { ReviewJob } from '@/frameworks/queue/pQueueAdapter.js';
-import type { ReviewRequestTrackingGateway, Platform } from '@/interface-adapters/gateways/reviewRequestTracking.gateway.js';
+import type { ReviewRequestTrackingGateway, Platform } from '@/modules/tracking/interface-adapters/gateways/reviewRequestTracking.gateway.js';
 import type { Language } from '@/modules/shared-kernel/entities/language/language.schema.js';
 
 export interface TriggerReviewParams {
